Add Season & Patch Game Date Bug
Patched game date bug where I had the col Year that was effectively season but didnt account for the fact that each season spans across 2 years. This would have and was messing up with sorting games in the correct order for each season. Game link is also a dist/unique join key. Note that currently in the 1x2 odds file that Game_ID is an outdated and no longer useful col.